Move all test helpers to be together
