+++ src/core_tests/mod.rs
@@ -1,20 +1,22 @@
-mod activity_tasks;
-mod queries;
-mod replay_flag;
-mod workers;
-mod workflow_tasks;
+use std::time::Duration;
 
+use futures::FutureExt;
+use tokio::time::sleep;
+
+use crate::test_help::{build_fake_core, fake_sg_opts, hist_to_poll_resp, TEST_Q};
 use crate::{
     errors::{PollActivityError, PollWfError},
-    machines::test_help::{build_fake_core, fake_sg_opts, hist_to_poll_resp, TEST_Q},
     pollers::MockManualGateway,
     protos::temporal::api::workflowservice::v1::PollActivityTaskQueueResponse,
     test_help::canned_histories,
     Core, CoreInitOptionsBuilder, CoreSDK, WorkerConfigBuilder,
 };
-use futures::FutureExt;
-use std::time::Duration;
-use tokio::time::sleep;
+
+mod activity_tasks;
+mod queries;
+mod replay_flag;
+mod workers;
+mod workflow_tasks;
 
 #[tokio::test]
 async fn after_shutdown_server_is_not_polled() {
