+++ src/test_help/history_info.rs
@@ -1,7 +1,7 @@
-use crate::workflow::HistoryUpdate;
 use crate::{
     protos::temporal::api::enums::v1::EventType,
     protos::temporal::api::history::v1::{History, HistoryEvent},
+    workflow::HistoryUpdate,
 };
 
 #[derive(Clone, Debug, PartialEq)]
